+++ src/components/mission-scheduling.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <div class="row items-center q-gutter-x-md">
+    <div class="row items-center q-gutter-x-lg">
       <div class="text-bold">
         Scheduling:
       </div>
@@ -9,16 +9,16 @@
         label="ASAP"
         val="ASAP"
         dense
-        :value="schedType"
-        @input="val => { $emit('schedType', val) }"
+        :value="schedInfo.schedType"
+        @input="schedType => emit({schedType})"
       />
 
       <q-radio
         label="Queue"
         val="QUEUE"
         dense
-        :value="schedType"
-        @input="val => { $emit('schedType', val) }"
+        :value="schedInfo.schedType"
+        @input="schedType => emit({schedType})"
       />
 
       <div class="row">
@@ -26,39 +26,51 @@
           label="At:"
           val="DATE"
           dense
-          :value="schedType"
-          @input="val => { $emit('schedType', val) }"
+          :value="schedInfo.schedType"
+          @input="schedType => emit({schedType})"
         />
-
-        <q-input
-          :disable="schedType !== 'DATE'"
-          class="q-pl-sm q-pr-sm"
-          style="width:13em"
-          dense autofocus
-          v-model="dateStr"
+        <q-chip
+          square
+          class="q-ml-sm bg-blue-1"
+          style="min-width:12em"
         >
-          <template v-slot:append>
-            <q-icon name="event" class="cursor-pointer" size="20px">
-              <q-popup-proxy transition-show="scale" transition-hide="scale">
-                <q-date
-                  v-model="dateStr"
-                  :mask="mask"
-                  today-btn
-                />
-              </q-popup-proxy>
-            </q-icon>
-            <q-icon name="access_time" class="cursor-pointer" size="20px">
-              <q-popup-proxy transition-show="scale" transition-hide="scale">
-                <q-time
-                  v-model="dateStr"
-                  :mask="mask"
-                  format24h
-                  now-btn
-                />
-              </q-popup-proxy>
-            </q-icon>
-          </template>
-        </q-input>
+          {{ schedInfo.schedDate }}
+        </q-chip>
+        <q-popup-edit
+          v-model="dateStr"
+          title="Mission Template ID"
+          buttons
+          @before-show="setDateStr"
+          @save="setSchedDate"
+        >
+          <div class="column">
+            <div class="row justify-cetner">
+              <q-input
+                :disable="schedInfo.schedType !== 'DATE'"
+                class="q-pl-lg q-pr-lg bg-blue-1"
+                :style="schedInfo.schedType === 'DATE' ? 'width:13em' : 'width:3em'"
+                dense autofocus borderless
+                v-model="dateStr"
+              />
+            </div>
+          </div>
+          <div class="row q-gutter-x-xs">
+            <q-date
+              v-model="dateStr"
+              :mask="mask"
+              flat
+              today-btn
+            />
+            <q-time
+              v-model="dateStr"
+              :mask="mask"
+              format24h
+              flat
+              now-btn
+            />
+          </div>
+        </q-popup-edit>
+
       </div>
     </div>
   </div>
@@ -73,14 +85,9 @@
     name: 'mission-scheduling',
 
     props: {
-      schedType: {
-        type: String,
+      schedInfo: {
+        type: Object,
         required: true
-      },
-
-      date: {
-        type: Date,
-        required: false
       },
     },
 
@@ -94,14 +101,38 @@
       dateStr: '',
     }),
 
-    mounted() {
-      this.dateStr = this.date && date.formatDate(this.date, this.mask)
-    },
+    methods: {
+      setDateStr() {
+        this.dateStr = ''
+        const {schedType, schedDate} = this.schedInfo
+        if (schedType === 'DATE') {
+          const theDate = schedDate ? new Date(schedDate) : new Date()
+          this.dateStr = date.formatDate(theDate, this.mask)
+        }
+      },
 
-    watch: {
-      dateStr(val) {
-        const schedDate = date.extractDate(val, this.mask)
-        this.$emit('schedDate', schedDate)
+      setSchedDate(val) {
+        const schedDate = val ? new Date(date.extractDate(val, this.mask)).toISOString() : null
+        this.emit({schedDate})
+      },
+
+      emit(patch) {
+        const schedInfo = {... this.schedInfo}
+        if (patch.schedDate) {
+          schedInfo.schedDate = patch.schedDate
+        }
+        if (patch.schedType) {
+          schedInfo.schedType = patch.schedType
+        }
+        if (schedInfo.schedType === 'DATE') {
+          if (!schedInfo.schedDate) {
+            schedInfo.schedDate = date.formatDate(new Date(), this.mask)
+          }
+        }
+        else {
+          schedInfo.schedDate = null
+        }
+        this.$emit('schedInfo', schedInfo)
       },
     },
  }